+++ Application/EdFi.Ods.Admin.Api/Docker/mssql/run.sh
@@ -16,7 +16,9 @@ jq --arg variable "$measurementId" '.AppSettings.GoogleAnalyticsMeasurementId = 
 
 mv /app/temp.json /app/appsettings.json
 
-cp /ssl/server.crt /usr/local/share/ca-certificates/
-update-ca-certificates
+if [[ -f /ssl/server.crt ]]; then
+  cp /ssl/server.crt /usr/local/share/ca-certificates/
+  update-ca-certificates
+fi
 
 dotnet EdFi.Ods.Admin.Api.dll
